+++ src/view/trip-point.js
@@ -1,7 +1,7 @@
-import { formatDateOnly, formatTimeOnly, formatShortDate, formatFullDate, calculateDuration } from '../util.js';
+import { formatDateOnly, formatTimeOnly, formatShortDate, formatFullDate, calculateDuration } from '../util/date-format.js';
 
-export const createTripPoints = (createMockPoints) => {
-  const {type, city, dateFrom, dateTill, offers, price, isFavorite} = createMockPoints;
+export const createTripPoint = (createMockPoints) => {
+  const {type, destination, dateFrom, dateTill, offers, price, isFavorite} = createMockPoints;
 
   const createOffers = offers.map((offer) => {
     return `<li class="event__offer">
@@ -21,7 +21,7 @@ export const createTripPoints = (createMockPoints) => {
     </div>
 
 
-    <h3 class="event__title">${type} ${city}</h3>
+    <h3 class="event__title">${type} ${destination.city}</h3>
     <div class="event__schedule">
       <p class="event__time">
         <time class="event__start-time" datetime="${formatFullDate(dateFrom)}">${formatTimeOnly(dateFrom)}</time>
@@ -58,16 +58,3 @@ export const createTripPoints = (createMockPoints) => {
   </div>
 </li>`;
 };
-
-
-// /**
-//  * calculateDuration = (start, end)
-//  * Функция считает разницу дат (времени)
-//  * @param {date} start - начало промежутка
-//  * @param {date} end - конец промежутка
-//  * @returns {string} - возвращает промежуток в формате ${hours}H ${minutes}M
-//  * до 60 мин - возвращает ${minutes}M
-//  * // Math.round() возвращает число, округлённое к ближайшему целому.
-//  * // Math.floor() - округление вниз до ближайшего меньшего целого.
-//  * // % - возвращает целочисленный остаток от деления двух операндов.
-//  */
